Ivan Yesilevskyi 1. Moved user checking from service to controller

+++ si-webapp/src/main/java/com/shutafin/controller/UserImageController.java
@@ -27,6 +27,9 @@ public class UserImageController {
     public UserImageWeb getUserImage(@RequestHeader(value = "session_id") String sessionId,
                                      @RequestParam(value = "image_id") Long userImageId) {
         User user = sessionManagementService.findUserWithValidSession(sessionId);
+        if (user == null) {
+            throw new AuthenticationException();
+        }
         UserImage image = userImageService.getUserImage(user, userImageId);
         return new UserImageWeb(image.getImageStorage().getImageEncoded(), image.getCreatedDate().toString());
     }
@@ -35,6 +38,9 @@ public class UserImageController {
     public void addUserImage(@RequestHeader(value = "session_id") String sessionId,
                                @RequestBody UserImageWeb image) {
         User user = sessionManagementService.findUserWithValidSession(sessionId);
+        if (user == null) {
+            throw new AuthenticationException();
+        }
         userImageService.addUserImage(image, user);
     }
 
@@ -42,6 +48,9 @@ public class UserImageController {
     public void deleteUserImage(@RequestHeader(value = "session_id") String sessionId,
                                @RequestParam(value = "image_id") Long userImageId) {
         User user = sessionManagementService.findUserWithValidSession(sessionId);
+        if (user == null) {
+            throw new AuthenticationException();
+        }
         UserImage image = userImageService.getUserImage(user, userImageId);
         userImageService.deleteUserImage(image);
     }
